+++ scene/wlf_scene_tree.c
@@ -88,7 +88,7 @@ static struct wlf_scene_node *scene_node_at(struct wlf_scene_node *node,
 }
 
 static void scene_node_bounds(struct wlf_scene_node *node,
-		int x, int y, struct wlf_region *visible) {
+		double x, double y, struct wlf_region *visible) {
 	if (!node->state.enabled) {
 		return;
 	}
